testTransferTx

+++ db/migration/000001_init_schema.up.sql
@@ -9,7 +9,7 @@ CREATE TABLE "accounts" (
 CREATE TABLE "entries" (
   "id" bigserial PRIMARY KEY,
   "account_id" bigint NOT NULL,
-  "amount" bigint NOT NULL,
+  "amount" decimal NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT (now())
 );
 
@@ -17,7 +17,7 @@ CREATE TABLE "transfers" (
   "id" bigserial PRIMARY KEY,
   "from_account_id" bigint NOT NULL,
   "to_account_id" bigint NOT NULL,
-  "amount" bigint NOT NULL,
+  "amount" decimal NOT NULL,
   "created_at" timestamp NOT NULL DEFAULT (now())
 );
